Update bitcoind_install.sh
Tidied a couple of things up - you don't need apt-get upgrade as well as apt-get dist-upgrade.

Also... I've removed all of the sudos - and run the whole thing as root - there's too many restricted commands to run as anything else! I've added an if at the beginning of the script to check if you are root and to exit if not.